Patch 0157
・script.cのbuildinで使われている変数の一部をshort型からint型に変更（script.c, unit.*）
・スクリプトで使える基数として2進数をサポート（script.c）
	0bを頭に付けて記述します。
	ビットフラグを明瞭に表現したいときなどに。
	0b1101 [2進数] = 13 [10進数] = 0xd [16進数] = 015 [8進数]

+++ src/map/unit.c
@@ -1790,7 +1790,7 @@ int unit_mobstopattacked(struct map_session_data *sd,va_list ap)
  * 見た目のサイズを変更する
  *------------------------------------------
  */
-int unit_changeviewsize(struct block_list *bl,short size)
+int unit_changeviewsize(struct block_list *bl,int size)
 {
 	nullpo_retr(0, bl);
