+++ src/map/unit.h
@@ -63,7 +63,7 @@ struct unit_data* unit_bl2ud(struct block_list *bl);
 int unit_remove_map(struct block_list *bl, int clrtype, int flag);
 int unit_distance(int x0,int y0,int x1,int y1);
 int unit_free(struct block_list *bl, int clrtype);
-int unit_changeviewsize(struct block_list *bl,short size);
+int unit_changeviewsize(struct block_list *bl,int size);
 
 // 初期化ルーチン
 int do_init_unit(void);
